+++ src/lib/NodeSDK.ts
@@ -212,9 +212,9 @@ class NodeSDK {
      *                          DataStoreType.NoPermanentStore Tell the server to NOT store the messages for history of the bot and the contact. But being stored temporarily as a normal part of delivery (e.g. if the recipient is offline at the time of sending).<br>
      *                          DataStoreType.StoreTwinSide The messages are fully stored.<br>
      *                          DataStoreType.UsestoreMessagesField to follow the storeMessages SDK's parameter behaviour.<br>
-     *       "autoInitialBubblePresence" to allow automatic opening of conversation to the bubbles with sending XMPP initial presence to the room. Default value is true.
-     *       "autoLoadConversations" to activate the retrieve of conversations from the server. The default value is true.
-     *       "autoLoadContacts" to activate the retrieve of contacts from roster from the server. The default value is true.
+     *       "autoInitialBubblePresence" to allow automatic opening of conversation to the bubbles with sending XMPP initial presence to the room. Default value is true. </br>
+     *       "autoLoadConversations" to activate the retrieve of conversations from the server. The default value is true. </br>
+     *       "autoLoadContacts" to activate the retrieve of contacts from roster from the server. The default value is true. </br>
      *   },<br>
      *   // Services to start. This allows to start the SDK with restricted number of services, so there are less call to API.<br>
      *   // Take care, severals services are linked, so disabling a service can disturb an other one.<br>
@@ -251,12 +251,12 @@ class NodeSDK {
     constructor(options) {
         /*
              *       @ deprecated "storeMessages": false, Tell the server to store the message for delay distribution and also for history. Please avoid to set it to true for a bot which will not read anymore the messages. It is a better way to store it in your own CPaaS application<br>
-     *       "nbMaxConversations": 15, Parameter to set the maximum number of conversations to keep (defaut value to 15). Old ones are remove from XMPP server with the new method `ConversationsService::removeOlderConversations`.
-     *       "rateLimitPerHour": 1000, Parameter to set the maximum of "message" stanza sent to server by hour. Default value is 1000.
-     *       "messagesDataStore": DataStoreType.NoStoreBotSide, Parameter to define the behaviour of the storage of the messages (Enum DataStoreType in lib/config/config , default value "DataStoreType.NoStoreBotSide")
-     *                          DataStoreType.NoStore Same behaviour as previously `storeMessages=false` Tell the server to NOT store the messages for delay distribution or for history of the bot and the contact.
-     *                          DataStoreType.NoStoreBotSide The messages are not stored on  loggued-in Bot's history, but are stored on the other side. So the contact kept the messages exchanged with bot in his history.
-     *                          DataStoreType.StoreTwinSide The messages are fully stored.
+     *       "nbMaxConversations": 15, Parameter to set the maximum number of conversations to keep (defaut value to 15). Old ones are remove from XMPP server with the new method `ConversationsService::removeOlderConversations`.</br>
+     *       "rateLimitPerHour": 1000, Parameter to set the maximum of "message" stanza sent to server by hour. Default value is 1000.</br>
+     *       "messagesDataStore": DataStoreType.NoStoreBotSide, Parameter to define the behaviour of the storage of the messages (Enum DataStoreType in lib/config/config , default value "DataStoreType.NoStoreBotSide")</br>
+     *                          DataStoreType.NoStore Same behaviour as previously `storeMessages=false` Tell the server to NOT store the messages for delay distribution or for history of the bot and the contact.</br>
+     *                          DataStoreType.NoStoreBotSide The messages are not stored on  loggued-in Bot's history, but are stored on the other side. So the contact kept the messages exchanged with bot in his history.</br>
+     *                          DataStoreType.StoreTwinSide The messages are fully stored.</br>
 
          */
         /* process.on("uncaughtException", (err) => {
@@ -296,10 +296,10 @@ class NodeSDK {
      * @instance
      * @param {String} token a valid token to login without login/password.
      * @description
-     *    Start the SDK
-     *    Note :
-     *    The token must be empty to signin with credentials.
-     *    The SDK is disconnected when the renew of the token had expired (No initial signin possible with out credentials.)
+     *    Start the SDK </br>
+     *    Note :</br>
+     *    The token must be empty to signin with credentials.</br>
+     *    The SDK is disconnected when the renew of the token had expired (No initial signin possible with out credentials.)</br>
      * @memberof NodeSDK
      */
     start(token) {
@@ -733,20 +733,20 @@ class NodeSDK {
      * @method getConnectionStatus
      * @instance
      * @description
-     *    Get connections status of each low layer services, and also the full SDK state.
-     *
-     * {
-     * restStatus: boolean, The status of the REST connection authentication to rainbow server.
-     * xmppStatus: boolean, The status of the XMPP Connection to rainbow server.
-     * s2sStatus: boolean, The status of the S2S Connection to rainbow server.
-     * state: SDKSTATUSENUM The state of the SDK.
-     * nbHttpAdded: number, the number of HTTP requests (any verb GET, HEAD, POST, ...) added in the HttpManager queue. Note that it is reset to zero when it reaches Number.MAX_SAFE_INTEGER value.
-     * httpQueueSize: number, the number of requests stored in the Queue. Note that when a request is sent to server, it is already removed from the queue.
-     * nbRunningReq: number, the number of requests which has been poped from the queue and the SDK did not yet received an answer for it.
-     * maxSimultaneousRequests : number, the number of request which can be launch at a same time.
-     * nbReqInQueue : number, the number of requests waiting for being treated by the HttpManager. 
-     * }
-     * @return {Promise<{restStatus: boolean, xmppStatus: boolean, s2sStatus: boolean, state: SDKSTATUSENUM, nbHttpAdded: number, httpQueueSize: number, nbRunningReq: number, maxSimultaneousRequests : number}>}
+     *    Get connections status of each low layer services, and also the full SDK state. </br>
+     * </br>
+     * { </br>
+     * restStatus: boolean, The status of the REST connection authentication to rainbow server. </br>
+     * xmppStatus: boolean, The status of the XMPP Connection to rainbow server. </br>
+     * s2sStatus: boolean, The status of the S2S Connection to rainbow server. </br>
+     * state: SDKSTATUSENUM The state of the SDK. </br>
+     * nbHttpAdded: number, the number of HTTP requests (any verb GET, HEAD, POST, ...) added in the HttpManager queue. Note that it is reset to zero when it reaches Number.MAX_SAFE_INTEGER value. </br>
+     * httpQueueSize: number, the number of requests stored in the Queue. Note that when a request is sent to server, it is already removed from the queue. </br>
+     * nbRunningReq: number, the number of requests which has been poped from the queue and the SDK did not yet received an answer for it. </br>
+     * maxSimultaneousRequests : number, the number of request which can be launch at a same time. </br>
+     * nbReqInQueue : number, the number of requests waiting for being treated by the HttpManager.  </br>
+     * } </br>
+     * @return {Promise<{ </br>restStatus: boolean, </br> xmppStatus: boolean, </br> s2sStatus: boolean, </br> state: SDKSTATUSENUM, </br> nbHttpAdded: number, </br> httpQueueSize: number, </br> nbRunningReq: number, </br> maxSimultaneousRequests : number </br>}>}
      * @category async
      */
     getConnectionStatus () : Promise<{restStatus: boolean, xmppStatus: boolean, s2sStatus: boolean, state: SDKSTATUSENUM, nbHttpAdded: number, httpQueueSize: number, nbRunningReq: number, maxSimultaneousRequests : number}> {